some refactor and edit profile handler success #744
