+++ WMIAdventure/frontend/src/js/api/gateways/user-profile-update/UserUtils.js
@@ -6,7 +6,7 @@ const createFormDataToSendAvatarImage = (basicUserData) => {
         formData.append('image', basicUserData.image, basicUserData.image.name);
 
     // Required data that has to be sent even if we only want to send image.
-    formData.append('user', basicUserData.user);
+    formData.append('user', basicUserData.userId);
     formData.append('displayedUsername', basicUserData.username);
     formData.append('semester', basicUserData.semester);
