+++ WMIAdventure/frontend/src/js/pages/Profile/styled-componets/RightDeckContainer.js
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 const RightDeckContainer = styled.div`
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
+  justify-content: center;
   align-items: center;
   width: calc(100% - 356px);
   height: 100%;
